New field proposal.
The proposed addition is a field named `versions` which
contains an object similar to `dependencies` in the `package.json`.
The project's dependencies would need to satisfy these restriction
for a mapping to be taken into account.

It might be more useful to provide not a simple JSON, but the already
filtered version (we could use a function or expose a js file that
does the hardwork for the user).